requirements.txt
asgiref==3.8.1
Django==5.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.1
whitenoise==6.9.0
langchain
langchain_community
langchain_huggingface
langchain-google-genai
langchain_core
openai
python-dotenv